+++ jdbc/src/ConfigDataSource.kt
@@ -3,9 +3,12 @@ package klite.jdbc
 import klite.Config
 import klite.info
 import klite.logger
+import klite.warn
 import java.io.PrintWriter
+import java.lang.Thread.sleep
 import java.sql.Connection
 import java.sql.DriverManager
+import java.sql.SQLException
 import java.sql.SQLFeatureNotSupportedException
 import java.util.*
 import javax.sql.DataSource
@@ -18,6 +21,8 @@ open class ConfigDataSource(
   val isReadOnly: Boolean = Config.optional("DB_READONLY") == "true",
   val props: Properties = Properties()
 ): DataSource {
+  private val log = logger()
+
   init {
     user?.let { props["user"] = it }
     pass?.let { props["password"] = it }
@@ -33,7 +38,19 @@ open class ConfigDataSource(
   }
 
   init {
-    logger().info("Connecting to $url${user?.let { ", user: $user" } ?: ""}")
+    log.info("Connecting to $url${user?.let { ", user: $user" } ?: ""}")
+  }
+
+  fun waitForAcceptConnections(numTries: Int = Config.optional("DB_WAIT_TRIES", "10").toInt(), timeoutMs: Long = 1000L) {
+    var tries = 1
+    do {
+      try { connection.use { return } }
+      catch (e: SQLException) {
+        log.warn("Retrying $tries: $e")
+        sleep(timeoutMs)
+      }
+    } while (++tries <= numTries)
+    error("Could not connect to $url after $numTries tries")
   }
 
   override fun getConnection(): Connection = DriverManager.getConnection(url, props)